backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
psycopg[binary]==3.1.19
argon2-cffi==23.1.0
python-multipart==0.0.9
